fix: extract shared plural rules, add daily completeness check, harden release
- Extract PLURAL_SUFFIXES to scripts/plural-rules.js (single source of truth
  for translate-locale.js, check-completeness.js, cleanup-stale-keys.js)
- Add km (Khmer) to plural suffixes and clean 4 stale _one keys
- Add daily scheduled workflow (.github/workflows/check-completeness.yml)
  that runs completeness check and opens a GitHub issue if errors > 0
- Add cleanup:stale and generate:fallbacks steps to release workflow
  so stale keys and outdated fallbacks are fixed before completeness gate
