+++ translations/js/km/monorepo/core.json
@@ -193,7 +193,6 @@
   "common.private": "ឯកជន",
   "common.pro": "Pro",
   "common.product": "ផលិតផល",
-  "common.product_found_locally_one": "រកឃើញផលិតផល {count} នៅក្នុងស្រុក",
   "common.product_found_locally_other": "រកឃើញផលិតផល {count} នៅក្នុងស្រុក",
   "common.product_name": "ឈ្មោះផលិតផល",
   "common.product_settings": "ការកំណត់ផលិតផល",
@@ -293,7 +292,6 @@
   "common.value": "តម្លៃ",
   "common.variants": "បំរែបំរួល",
   "common.variation": "បម្រែបម្រួល",
-  "common.variation_found_for_term_one": "រកឃើញ {count} បម្រែបម្រួលសម្រាប់ \"{term}\"",
   "common.variation_found_for_term_other": "រកឃើញ {count} បម្រែបម្រួលសម្រាប់ \"{term}\"",
   "common.version": "{platform} កំណែ",
   "common.view_all_tax_rates": "មើលអត្រាពន្ធទាំងអស់",
