+++ translations/js/km/woocommerce-pos-pro/wp-admin-store-edit.json
@@ -56,7 +56,6 @@
   "tax.wc_settings": "ការកំណត់ពន្ធ WooCommerce",
   "users.description": "ជ្រើសរើសអ្នកប្រើដែលអាចចូលដំណើរការហាងនេះនៅក្នុង POS។ ទុកឲ្យទទេដើម្បីអនុញ្ញាតឲ្យអ្នកប្រើ POS ទាំងអស់។",
   "users.no_users": "រកមិនឃើញអ្នកប្រើដែលមានការចូលដំណើរការនៅ POS។",
-  "users.selected_one": "បានជ្រើសរើសអ្នកប្រើ {count} នាក់",
   "users.selected_other": "បានជ្រើសរើសអ្នកប្រើ {count} នាក់",
   "users.title": "អ្នកប្រើដែលបានអនុញ្ញាត"
 }
