+++ translations/js/km/woocommerce-pos/wp-admin-settings.json
@@ -82,7 +82,6 @@
   "sessions.minutes_ago": "{minutes} នាទីមុន",
   "sessions.my_sessions": "សម័ង្គ្រាម របស់ខ្ញុំ",
   "sessions.no_active_sessions": "មិនមាន សម័ង្គ្រាម ដែលកំពុងដើរ",
-  "sessions.session_count_one": "{count} សម័យកាល",
   "sessions.session_count_other": "{count} សម័យកាល",
   "sessions.session_terminated": "សម័យកាលបានបញ្ចប់ដោយជោគជ័យ",
   "sessions.sessions": "សម័ង្គ្រាម",
